requirements.txt
# Core dependencies
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.22.3

# ML/AI dependencies
torch==2.0.1
transformers==4.30.2
huggingface-hub==0.19.4
accelerate==0.25.0
numpy==1.24.3

# Audio processing
TTS==0.19.0  # Updated from 0.17.3 to a version compatible with Python 3.9
librosa==0.10.0.post2
soundfile==0.12.1

# Frontend
gradio==3.50.2

# Testing
pytest==7.4.0

# Visualization
matplotlib==3.7.1

# Additional dependencies
pyrubberband==0.3.0